requirements.txt
pandas==1.4.3
numpy==1.22.4
tdqm==4.64.0
fonttools==4.37.1
opencv==4.6.0
pillow==9.2.0